PR-448199: Moving Server to src directory.
